finalizes the project for delivery and fix erros

+++ src/main/java/br/com/fernanda/springbootinterview/api/ClientAPI.java
@@ -3,7 +3,6 @@ package br.com.fernanda.springbootinterview.api;
 import br.com.fernanda.springbootinterview.exception.ResourceAlreadyRegisteredException;
 import br.com.fernanda.springbootinterview.exception.ResourceNotFoundException;
 import br.com.fernanda.springbootinterview.exception.InvalidArgumentException;
-import br.com.fernanda.springbootinterview.model.City;
 import br.com.fernanda.springbootinterview.service.CityService;
 import br.com.fernanda.springbootinterview.service.ClientService;
 import io.swagger.annotations.Api;
@@ -16,7 +15,11 @@ import br.com.fernanda.springbootinterview.util.DateUtil;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import javax.transaction.Transactional;
 import javax.validation.Valid;
+import javax.websocket.server.PathParam;
+
 import lombok.Setter;
 
 
@@ -26,7 +29,7 @@ import lombok.Setter;
 @Api(value = "Client API")
 public class ClientAPI {
 
-    private static final String MESSAGE_CLIENT_REGISTERED = "The client cannot be registered ";
+    private static final String MESSAGE_CLIENT_REGISTERED = "Client already registered ";
     private static final String MESSAGE_CLIENT_NOT_FOUND = "Client not found";
     private static final String MESSAGE_INVALID_GENDER = "Gender must be 'F' or 'M' in upper case.";
     private static final String MESSAGE_INVALID_BIRTH_DATE = "Invalid birth date. " +
@@ -34,14 +37,15 @@ public class ClientAPI {
 
     @Autowired
     private ClientService clientService;
+
     @Autowired
     private CityService cityService;
 
     @GetMapping("/findByName/{name}")
     @ApiOperation(value = "** FIND CLIENT BY NAME **")
     public ResponseEntity<?> findByName( @PathVariable("name") String name) {
-        Client client = clientService.findByName(name.toUpperCase());
-        if(client == null ){
+        Client client = this.clientService.findByName(name);
+        if (client == null){
             throw new ResourceNotFoundException(MESSAGE_CLIENT_NOT_FOUND);
         }
         ClientDTO clientDTO = new ClientDTO();
@@ -52,72 +56,78 @@ public class ClientAPI {
     @GetMapping("/findById/{id}")
     @ApiOperation(value = "** FIND CLIENT BY ID **")
     public ResponseEntity<?> findById( @PathVariable("id") long id) {
-        this.verifyIfClientExistsById(id);
-        Client client = clientService.findById(id);
-        ClientDTO clientDTO =  this.clientDTOMapper(client, new ClientDTO());
+        Client client = this.clientService.findById(id);
+        if (client == null){
+            throw new ResourceNotFoundException(MESSAGE_CLIENT_NOT_FOUND);
+        }
+
+        ClientDTO clientDTO = new ClientDTO();
+        clientDTO = this.clientDTOMapper(client, clientDTO);
         return new ResponseEntity<>(clientDTO, HttpStatus.OK);
     }
 
-    @DeleteMapping("/remove")
+    @DeleteMapping("/remove/{id}")
     @ApiOperation(value = "** REMOVE CLIENT **")
-    public ResponseEntity<?> remove( @RequestBody ClientDTO clientDTO) {
-        this.verifyIfClientExistsByName(clientDTO.getName());
-        Client client = clientService.findByName(clientDTO.getName());
+    @Transactional
+    public ResponseEntity<?> remove(@PathVariable("id") long id) {
+
+        Client client = this.clientService.findById(id);
+        if (client == null){
+            throw new ResourceNotFoundException(MESSAGE_CLIENT_NOT_FOUND);
+        }
         this.clientService.remove(client);
-        return new ResponseEntity(clientDTO, HttpStatus.OK);
+        ClientDTO clientDTO = new ClientDTO();
+        clientDTO = this.clientDTOMapper(client, clientDTO);
+        return new ResponseEntity<>(clientDTO, HttpStatus.OK);
     }
 
     @PostMapping("/save")
     @ApiOperation(value = "** SAVE CLIENT **")
     public ResponseEntity<?> save( @Valid @RequestBody ClientDTO clientDTO) {
-        Client client = new Client();
-        City city = this.cityService.validateCity(clientDTO.getCity());
+        this.cityService.validateCity(clientDTO.getCity());
 
-        if(!verifyIfClientExistsByName(clientDTO.getName())){
+        Client client = this.clientService.findByName(clientDTO.getName());
+
+        if (client == null){
             if(DateUtil.isValidDate(clientDTO.getBirthDate())){
+               client = new Client();
                client = this.clientMapper(clientDTO, client);
-               client.getCity().setId(city.getId());
                client.setId(null);
                this.clientService.save(client);
             }else{
                 throw new InvalidArgumentException(MESSAGE_INVALID_BIRTH_DATE);
             }
-        }
-        else{
+        }else{
             throw new ResourceAlreadyRegisteredException(MESSAGE_CLIENT_REGISTERED);
         }
 
-        clientDTO.setId(client.getId());
+        Client client2 = this.clientService.findByName(clientDTO.getName());
+        clientDTO = this.clientDTOMapper(client2, clientDTO);
         return new ResponseEntity<>(clientDTO, HttpStatus.CREATED);
     }
 
-    @PutMapping("/updateName")
+    @PutMapping("/updateName/{id}")
     @ApiOperation(value = "** UPDATE CLIENT NAME **")
-    public ResponseEntity<?> updateName( @RequestBody ClientDTO clientDTO) {
-        this.verifyIfClientExistsById(clientDTO.getId());
+    @Transactional
+    public ResponseEntity<?> updateName( @PathVariable("id") long id, @RequestBody String name) {
 
-        Client client = this.clientService.findById(clientDTO.getId());
-        client.setName(clientDTO.getName().toUpperCase());
+        Client client = this.clientService.findById(id);
+
+        if (client == null){
+            throw new ResourceNotFoundException(MESSAGE_CLIENT_NOT_FOUND);
+        }
+
+        client.setName(name);
         this.clientService.save(client);
 
+        ClientDTO clientDTO = new ClientDTO();
         clientDTO = this.clientDTOMapper(client, clientDTO);
         return new ResponseEntity<>(clientDTO, HttpStatus.OK);
     }
 
 
-    private boolean verifyIfClientExistsByName(String name){
-        if (this.clientService.findByName(name.toUpperCase()) == null){
-            return false;
-        }
-        return true;
-    }
-
-    private void verifyIfClientExistsById(Long id){
-        if (this.clientService.findById(id) == null)
-            throw new ResourceNotFoundException(MESSAGE_CLIENT_NOT_FOUND);
-    }
-
     private Client clientMapper(ClientDTO clientDTO, Client client) {
+
         client.setName(clientDTO.getName().toUpperCase());
 
         if (clientDTO.getGender() == 'M') {
@@ -132,12 +142,13 @@ public class ClientAPI {
        client.setBirthDate(DateUtil.convertStringToLocalDate(clientDTO.getBirthDate()));
        client.setAge(clientDTO.getAge());
        client.setId(clientDTO.getId());
-       
+
        return client;
        
     }
 
     private ClientDTO clientDTOMapper(Client client, ClientDTO clientDTO){
+
         clientDTO.setName(client.getName());
         clientDTO.setGender(client.getGender().getValue());
         clientDTO.setCity(client.getCity());
@@ -148,5 +159,6 @@ public class ClientAPI {
         return clientDTO;
 
     }
+
 
 }
